requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
boto3==1.37.15
botocore==1.37.38
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
coverage==4.4.1
cryptography==44.0.0
datedelta==1.4
DBUtils==1.3
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
django-rest-marshmallow==4.0.2
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.0
elasticsearch==7.17.0
email-validator==2.1.0
environs==11.2.1
executing==2.1.0
fastapi==0.115.6
fastapi_cors==0.0.6
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google_search_results==2.4.2
googleapis-common-protos==1.69.2
googlemaps==4.10.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.2.0
multipledispatch==1.0.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.24.3
openai==1.3.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pdfkit==1.0.0
pexpect==4.9.0
pip-tools==7.4.1
platformdirs==3.11.0
polyline==2.0.2
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydyf==0.3.0
Pygments==2.19.1
PyMySQL==0.10.0
pyparsing==2.4.7
pypdf==3.17.2
PyPDF2==1.26.0
pyphen==0.13.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.6
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2020.1
PyYAML==5.3.1
pyzmq==26.2.0
razorpay==1.4.1
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.11.5
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.3
sentry-sdk==1.9.0
shapely==2.0.4
simplejson==3.17.2
six==1.15.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.1.1
tokenizers==0.15.2
tomli==2.2.1
torch==2.1.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.36.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.0
tzdata==2022.5
uri-template==1.3.0
uritemplate==3.0.1
uritools==4.0.2
urllib3==1.25.10
uv==0.7.3
uvicorn==0.34.0
virtualenv==20.24.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.5
widgetsnbextension==4.0.9
xlrd==2.0.1
XlsxWriter==1.3.7
xlutils==2.0.0
xlwt==1.3.0
yarl==1.18.3
zipp==3.1.0
zopfli==0.2.3
zstandard==0.23.0
